feat: responsive redesign of Jam pages
Remove inline styles across all Jam pages. Extract JamResultsPanel
from JamStudentView (452 -> 287 lines). Apply design system card
patterns and responsive modal/form layouts.
